\name{VotingGames-package}
\alias{VotingGames-package}
\alias{VotingGames}
\docType{package}
\title{
  Methods for coalitional games.
}
\description{
  Compute Shapley and Banzhaf indices of coalitional games.
  Contains advanced methods for Weighted voting games with integer weights.
  Also contains methods for analyzing gene expressions data.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Matyáš Křišťan

Maintainer: Matyáš Křišťan <matyas.kristan@gmail.com>
}
\references{
  TODO Link to the thesis.
}
\keyword{ coalitional games banzhaf shapley }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
